Update .gitignore
Updated .gitignore to ignore the .db file to allow for fresh builds with fresh databases. Also untracked the location where products are stored inside of the /Products/<product name> directory so this way the github will not hold any of the products that would not be in the database since we are not shipping it.