Created a config file
Made config file and switched most things over to using it.

Still need to update the colors though the entire program to be pulled from the config, and potentially clean up that theme section of the config for clarity.

Name Colors not Colours for universal clarity and potential integrations if needed later on even though it pains me greatly...