Linked styling to of all screens to the config.
Updated all of the gui.py to use styling from the config which is then parsed into individual mid section of this in utils.py before being applied throughout the entire of gui.py.

Still need to adjust the first time admin section to use the config,ini and possible make 2 stage initialization to ensure it can use this.