Decided to backroll the start_fullscreen feature from the config.ini
The start_fullscreen feature from the config.ini could lead to incompat due to start_max_windowed being there and we have the ability to activate fullscreen with the bound f11. Therefore it is being removed from the config.

May be reimplemented later on if it seems worth the time.