Updated the edit_products function + More
Updated the edit_products function to use the layout of the view products page shown in the store, adjusted it to make it dynamic in its updated viewing size and added a placeholder image to avoid issues with there being no image and instruct to add one.

Currently only tested to be working when a singular change is made. Rework in another branch for the save_edit function since that is handling more than one change poorly (not clearing the old image when updating to a new one + other changes.

Untested are the category changes and listed changes.

Some changes were made in the file_manager and database to try and resolve this issue however i am making a radical change to a different method of assessing changes more robustly and am going to branch this product feature to test that before merging it back in.

The changes to the utils are the changes in update of the style for the product_page and the edit_product page since the gui changed and needed to style the new parts based off the product_page.