Save & Edit reworking for products
Save and Edit rework for the products now allows the user to edit the product in the same view style as the user would see the product to allow for consistency that is required when a product already exists (helps to envision the product during the edit creation phase). The handling now will correctly update and move the files as required including regenerating the qr code when either the name or price have changed (and keeping the iamge if there is one), or updating the image on its own, or doing a combination of the above. And it also responds as expected for the editing of non qr code or image parts of the product.

This push was a long time coming due to working on other assignments and Christmas & New Years Break after being stuck on it back when I was working on it going around in loops.

Rearranged parts in the todo markdown file since some of those may not be covered yet for required functionality of the product for the submission deadline. Somewhat streamlining the features I will provide to ensure I meet the basic requirements before continuing work on advanced features.