Adjusted the formatting of the cart page
Adjusted the formatting of the cart page to match the prior store page. Swapped the message label to match with the store page too.

Also adjusted the ability to go back to the store page to ensure it will work for the admin user after visiting the cart and then going back to the store page.