Many Changes, See notes for info:
Created a new scrollable grid to allow for use in the manage users and potentially manage categories and cart.

Updated the entire f the show_manage_users_screen to use grid to avoid conflicts with the scrollable grid frame.

Updated the display_users function to use grid.

Got the display users sections to properly fit under the headers (had to remove the width of the scrollbar (10) from the created width of the headers to make it align properly.

Updated the open_edit_dialong to properly have all required sections.

Removed the back to admin dashboard button in the dropdown while on the admin dashboard.

Updated the validation in the "open_edit_dialog" function

added a new function for all users to be able to edit themselves from the dropdown with proper validations. this is also in the form on a new popup window that stays over any other window the same as the edit window from the manage users function on the admin dashboard.

Adapted validation feature to be used for more cases of user validations.

Switched over all locations to use this new validation

Adjusted the "switch_to_change_password" function to be able to be called differently from login, self or admin. along with passing back to the prior dialog window if there was one open.

added the ability to manage self from dropdown for normal users and admins.

Swapped the is admin checkbox from manage users to be a dropdownbox instead