Many Changes see below for details:
Added new DB for coupons to be stored in,
Created new coupon management page for admins,
Created supporting functions in database.py and file_manager.py

Implemented the feature for coupons into the cart page.

Discount Manager allows the:
Creation,
Editing,
Toggling of state,
and deletion of coupons (as qr codes)

When using a qr code in the cart you can use either camera scanning or you can use the uploading of the image itself. This is then verified to ensure that the qr code is active and applies the discount to the total.

It will also track the uses of the qr code which can be seen in the management screen.