Implementing logging for ALL user and admin interractions.
Created the logging page on admin.
Created 2 tables one for UserActions and one for AdminActions.
implemented 3 functions into database.py for logging of user action, admin action and exporting the logs to a temp file.
Created a function in utils.py for logging of the actions. including the helper definitions of actions to call upon to simplify the implementation in gui.py

Adjusted the register message response for the auth.py function to return the same way as other methods and implemented its logging working.

Updated the "update_quantity" function in the cart to allow for logging changes in quantities for products made or removals from the cart.

All User logging functionality added to the gui.py, Testing proceeded that we needed to change the checkbox to a combobox for choosing to or not to log user actions, along with this we had to adjust the radioboxes to use a combobox too for the same error.

We also have had to change the combobox for logging use actions to be in the config.ini file. This is now working as expected.

added logging for attempts to log into an admin account.

adding logging for the success of each admin feature, and failure (failure to allow management to track failures to allow for extra top-up training).

Tested ALL logging features to ensure that they are working properly and that they have not broken anything else.

Storing TEMP log files inside of the local path directory for the product instead of the local directory of the computer to ensure that there aren't files all over the place. Management/cleanup of this activated on any closure or leaving of the logging page.