populating the dashboard page of the admin dashboard 1/3
Adding functionality to the currently blank admin dashboard page since we have completed the backlog of required features and are doing "polishing touches". This is a grid based layout which will allow for basic everyday monitoring that an administrative user can leave open for monitoring of the entire store page.

First off is implementing the admin logging to be displayed in this page but in smaller size for basic logging functions, whereas the logging page is more specifically for investigating actions made on a larger scale. This smaller display of it would allow the admin to be alerted to administrative failures and allow for investigation on the system logs page.

Still need to populate top left and top right on the dashboard.