populating the dashboard page of the admin dashboard 2/3
Populated the top left section of the admin dashboard with the "System Statistics" section which currently contains a count that updates of the "Total Products", "Listed Products", "Total Users" (excluding admin accounts) and "Active Discounts". This might get extended with more information and could potentially be transformed into a graphic system chart or of sorts if it seems like it'd fit better and be better for administration usage. Currently the number of 4 system statistics works great and fits perfectly when the dashboard window is shrank to its minimum size, however looks quite empty taking up less than half the height of the frame when on its default size.