populating the dashboard page of the admin dashboard 3/3
Populated the top Right section of the dashboard with the alerts section, including adding all of the supporting functions and adjusting the code so that it works properly.

Adjusted the login function to log failed admin attempts in both users and admins databases for logging.

Logging alerts for failed admin logins will go off if there are 2 or greater failed admin login attempts in under 30 minutes meaning it is more sensitive, whereas it will alert if a user account has 3 or greater login attempts that failed in 1 hour.

It will also alert on a stock being low if there are less than 5 stock on a product,

It will also now alert if there are 10 or more usages of coupon codes in an hour or less

Admin alert works, still needs to be tested with the rest of the alerts to ensure there are no errors.

Commented out old debugging statements added for logging implementation.