Update todo list post dashboard
Update the todolist to reflect changes in this branch, and add in the requirement to do testing of the application and redocument it.

Large change todo.md will now be up to date and reorganized properly considering potential changes and new features separated. and clearing of old fixed potential issues.