Improved handling for add and update of a product
Product adding and updating now allow for the removal of an image without the addition of another.
Image when removed from an updating product will handle it being null again and handles db and file's as expected.

When editing a product if listed it will check that no requirements have been removed, if they have you get a popup box to let you know what has been removed that is needed, and it will automatically take the product off being listed.

Resolved issue with qr code old not being cleaned up on update of product.